Refactor splitting zip files
When directory is given as input, simply pack it all and publish.

When zip file is specified, it is repacked to create smaller archives.
The tool tries to find the actual maven-repository files. If that
succeeds, only those files are pushed (meaning any ancillary files are
removed, e.g. licences, examples).

Fixes: NEXUS-237