Update error handling for creating new product
Start the process by cloning the configuration repo. This should catch
cases where the default system-wide configuration does not work at all.
An error message is printed in this case telling the user where to put
proper configuration.

It is still possible the clone will work, the product will be created
and then pushing will fail (for example because the user lacks
permission). In this case the content to add to the configuration file
is still printed.